requirements.txt
flask
psycopg2-binary
pytest
requests
playwright
Pillow
